split engine core dtos into modules
